tomo/YPSI/requirements.txt
numpy==2.4.4
opencv-python-headless==4.13.0.92
pandas==2.3.3
Pillow==11.3.0
streamlit==1.49.0
streamlit-drawable-canvas-fix==0.9.8
